+++ json_serializable/test/integration/json_test_example.dart
@@ -45,7 +45,7 @@ class Person {
       deepEquals(houseMap, other.houseMap);
 }
 
-@JsonSerializable()
+@JsonSerializable(constructor: 'custom')
 class Order {
   /// Used to test that `disallowNullValues: true` forces `includeIfNull: false`
   @JsonKey(disallowNullValue: true)
@@ -81,7 +81,7 @@ class Order {
   @JsonKey(ignore: true)
   bool? shouldBeCached;
 
-  Order(this.category, [Iterable<Item>? items])
+  Order.custom(this.category, [Iterable<Item>? items])
       : items = UnmodifiableListView<Item>(
             List<Item>.unmodifiable(items ?? const <Item>[]));
 
@@ -217,4 +217,25 @@ class UnknownEnumValue {
 
   factory UnknownEnumValue.fromJson(Map<String, dynamic> json) =>
       _$UnknownEnumValueFromJson(json);
+}
+
+@JsonSerializable(constructor: '_')
+class PrivateConstructor {
+  static int _id = 0;
+
+  final int id;
+  final String value;
+
+  PrivateConstructor._(this.id, this.value);
+
+  PrivateConstructor(this.value) : id = _id++;
+
+  factory PrivateConstructor.fromJson(Map<String, dynamic> json) =>
+      _$PrivateConstructorFromJson(json);
+
+  Map<String, dynamic> toJson() => _$PrivateConstructorToJson(this);
+
+  @override
+  bool operator ==(Object other) =>
+      other is PrivateConstructor && id == other.id && value == other.value;
 }