+++ json_serializable/test/json_serializable_test.dart
@@ -120,4 +120,5 @@ const _expectedAnnotatedTests = {
   'ValidToFromFuncClassStatic',
   'WithANonCtorGetter',
   'WithANonCtorGetterChecked',
+  'WrongConstructorNameClass',
 };